HTTP repository moved to core
