+++ documentation/src/main/cheatsheet/cheatsheets.adoc
@@ -98,6 +98,32 @@ Set format of access log. Default is <strong>LoggerFormat.DEFAULT</strong>.
 Configure when to collect access log. True if after request arrives, false after sending response.
 +++
 |===
+[[ClientDestination]]
+== ClientDestination
+
+
+[cols=">25%,^25%,50%"]
+[frame="topbot"]
+|===
+^|Name | Type ^| Description
+|[[domain]]`domain`|`String`|
++++
+Set the domain of the http repository
++++
+|[[hostHeader]]`hostHeader`|`String`|
++++
+Set the host header override to be used with a communication to the repository.
+ If it's set, it overrides any value in the 'Host' header, and sets the SNI SSL to the same value.
++++
+|[[port]]`port`|`Number (int)`|
++++
+Set the HTTP port of the http repository
++++
+|[[scheme]]`scheme`|`String`|
++++
+Set the URL scheme used to communicate with the repository
++++
+|===
 [[FilesystemRepositoryOptions]]
 == FilesystemRepositoryOptions
 
@@ -118,6 +144,41 @@ Set the EB address of the file system repository verticle
 +++
 Set the root folder of the repository on file system.
  If catalogue equals empty string a verticle will look for the files in classpath
++++
+|===
+[[HttpRepositoryOptions]]
+== HttpRepositoryOptions
+
+++++
+ Describes a configuration of Http Repository connector
+++++
+'''
+
+[cols=">25%,^25%,50%"]
+[frame="topbot"]
+|===
+^|Name | Type ^| Description
+|[[address]]`address`|`String`|
++++
+Set the EB address of the HTTP repository verticle
++++
+|[[allowedRequestHeaders]]`allowedRequestHeaders`|`Array of String`|
++++
+Set the collection of patterns of allowed request headers. Only headers matching any
+ of the pattern from the set will be sent to the HTTP repository
++++
+|[[clientDestination]]`clientDestination`|`link:cheatsheets.adoc#ClientDestination[ClientDestination]`|
++++
+Set the remote location of the repository
++++
+|[[clientOptions]]`clientOptions`|`link:cheatsheets.adoc#HttpClientOptions[HttpClientOptions]`|
++++
+Set the link used by the HTTP client
+ to communicate with remote http repository
++++
+|[[customHttpHeader]]`customHttpHeader`|`link:cheatsheets.adoc#CustomHttpHeader[CustomHttpHeader]`|
++++
+Set the header (name & value) to be sent in every request to the remote repository
 +++
 |===
 [[KnotxCSRFOptions]]
