+++ knotx-example/knotx-example-app/src/test/resources/knotx-test-app.json
@@ -1,7 +1,7 @@
 {
   "modules": [
     "server=io.knotx.server.KnotxServerVerticle",
-    "httpRepo=io.knotx.repository.HttpRepositoryConnectorVerticle",
+    "httpRepo=io.knotx.repository.http.HttpRepositoryConnectorVerticle",
     "fsRepo=io.knotx.repository.fs.FilesystemRepositoryConnectorVerticle",
     "splitter=io.knotx.knot.splitter.FragmentSplitterVerticle",
     "assembler=io.knotx.knot.assembler.FragmentAssemblerVerticle",
