+++ knotx-standalone/config/application.conf
@@ -1,6 +1,6 @@
 modules = [
     "server=io.knotx.server.KnotxServerVerticle"
-    "httpRepo=io.knotx.repository.HttpRepositoryConnectorVerticle"
+    "httpRepo=io.knotx.repository.http.HttpRepositoryConnectorVerticle"
     "fsRepo=io.knotx.repository.fs.FilesystemRepositoryConnectorVerticle"
     "splitter=io.knotx.knot.splitter.FragmentSplitterVerticle"
     "assembler=io.knotx.knot.assembler.FragmentAssemblerVerticle"
