+++ pom.xml
@@ -60,8 +60,7 @@
     <module>knotx-core</module>
     <module>knotx-junit</module>
     <module>knotx-mocks</module>
-
-    <module>knotx-repository-connector</module>
+    
     <module>knotx-adapter</module>
     <module>knotx-knot</module>
     <module>knotx-standalone</module>
