+++ tmp-core/pom.xml
@@ -118,11 +118,6 @@
       <artifactId>knotx-gateway</artifactId>
       <version>${project.version}</version>
     </dependency>
-    <dependency>
-      <groupId>io.knotx</groupId>
-      <artifactId>knotx-repository-connector-http</artifactId>
-      <version>${project.version}</version>
-    </dependency>
 
     <dependency>
       <groupId>org.jsoup</groupId>
