+++ tmp-core/src/main/java/io/knotx/repository/http/ClientDestination.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.knotx.repository;
+package io.knotx.repository.http;
 
 import io.vertx.codegen.annotations.DataObject;
 import io.vertx.core.json.JsonObject;
